requirements.txt
asgiref
Django~=3.2.12
djangorestframework==3.14.0
pillow
psycopg2-binary==2.9.9
pytz
sqlparse
django-filter~=23.5
djangorestframework-simplejwt
stripe==8.6.0